Health system complete
This branch has an independent health system that can be applied to any entity that can take damage.  There is a little green health bar over each entity and a large health bar on the game canvas.  Both are independent and can be applied to any scene.  There is also an end game menu that can exit or restart the game from the play mode not the build mode